+++ python/light-curve/light_curve/light_curve_py/features/otsusplit.py
@@ -6,8 +6,10 @@ from ._base import BaseFeature
 
 @dataclass()
 class OtsuSplit(BaseFeature):
-    """Difference of subset means, standard deviation of the lower subset, standard deviation
-    of the upper subset and upper-to-all observation count ratio for two subsets of magnitudes
+    """Otsu threshholding algorithm
+
+    Difference of subset means, standard deviation of the lower subset, standard deviation
+    of the upper subset and lower-to-all observation count ratio for two subsets of magnitudes
     obtained by Otsu's method split.
 
     Otsu's method is used to perform automatic thresholding. The algorithm returns a single
@@ -24,6 +26,7 @@ class OtsuSplit(BaseFeature):
     def _eval(self, t, m, sigma=None):
         n = len(m)
         amounts = np.arange(1, n)
+        m = np.sort(m)
 
         w0 = amounts / n
         w1 = 1 - w0
@@ -44,13 +47,13 @@ class OtsuSplit(BaseFeature):
         if len(m[arg + 1 :]) == 1:
             std_upper = 0
 
-        up_to_all_ratio = (arg + 1) / n
+        lower_to_all_ratio = (arg + 1) / n
 
-        return mean1[arg] - mean0[arg], std_lower, std_upper, up_to_all_ratio
+        return mean1[arg] - mean0[arg], std_lower, std_upper, lower_to_all_ratio
 
     @property
     def names(self):
-        return "otsu_mean_diff", "otsu_std_lower", "otsu_std_upper", "otsu_up_to_all_ratio"
+        return "otsu_mean_diff", "otsu_std_lower", "otsu_std_upper", "otsu_lower_to_all_ratio"
 
     @property
     def descriptions(self):
@@ -58,7 +61,7 @@ class OtsuSplit(BaseFeature):
             "difference between mean values of Otsu split subsets",
             "standard deviation for observations below the threshold given by Otsu method",
             "standard deviation for observations above the threshold given by Otsu method",
-            "ratio of quantity of observations above the threshold given by Otsu method to quantity of all observations",
+            "ratio of quantity of observations bellow the threshold given by Otsu method to quantity of all observations",
         )
 
     @property